extract and reuse experimentalAssets logic from `wrangler deploy` in `wrangler versions upload`
