+++ packages/wrangler/src/deploy/index.ts
@@ -1,9 +1,7 @@
-import { existsSync } from "node:fs";
 import path from "node:path";
 import { findWranglerToml, readConfig } from "../config";
 import { getEntry } from "../deployment-bundle/entry";
 import { UserError } from "../errors";
-import { getExperimentalAssetsBasePath } from "../experimental-assets";
 import {
 	getRules,
 	getScriptName,
@@ -16,6 +14,7 @@ import { getLegacyAssetPaths, getSiteAssetPaths } from "../sites";
 import { requireAuth } from "../user";
 import { collectKeyValues } from "../utils/collectKeyValues";
 import deploy from "./deploy";
+import { processExperimentalAssetsArg } from "./utils";
 import type { Config } from "../config";
 import type {
 	CommonYargsArgv,
@@ -290,38 +289,7 @@ export async function deployHandler(
 		);
 	}
 
-	const experimentalAssets = args.experimentalAssets
-		? { directory: args.experimentalAssets }
-		: config.experimental_assets;
-	if (experimentalAssets) {
-		const experimentalAssetsBasePath = getExperimentalAssetsBasePath(
-			config,
-			args.experimentalAssets
-		);
-		const resolvedExperimentalAssetsPath = path.resolve(
-			experimentalAssetsBasePath,
-			experimentalAssets.directory
-		);
-
-		if (!existsSync(resolvedExperimentalAssetsPath)) {
-			const sourceOfTruthMessage = args.experimentalAssets
-				? '"--experimental-assets" command line argument'
-				: '"experimental_assets.directory" field in your configuration file';
-
-			throw new UserError(
-				`The directory specified by the ${sourceOfTruthMessage} does not exist:\n` +
-					`${resolvedExperimentalAssetsPath}`
-			);
-		}
-
-		experimentalAssets.directory = resolvedExperimentalAssetsPath;
-	}
-
-	if (args.assets) {
-		logger.warn(
-			"The --assets argument is experimental and may change or break at any time"
-		);
-	}
+	const experimentalAssets = processExperimentalAssetsArg(args, config);
 
 	if (args.latest) {
 		logger.warn(
